server/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
langchain-text-splitters
langchain-classic
faiss-cpu
pypdf
sentence-transformers
python-dotenv
requests